analytics/pyproject.toml
[tool.poetry]
name = "analytics"
version = "0.1.0"
description = ""
authors = ["diana-create <dianamuse9@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
flask = "^3.0.3"
pymongo = "^4.10.1"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
flask-pymongo = "^2.3.0"
flask-cors = "^5.0.0"
ariadne = "0.23.0"
blinker = "^1.8.2"
click = "^8.1.7"
flask-sqlalchemy = "^3.1.1"
graphql-core = "^3.2.5"
idna = "^3.10"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.4"
markupsafe = "2.1.3"
packaging = "^24.1"
sniffio = "^1.3.1"
sqlalchemy = "^2.0.36"
starlette = "^0.41.0"
typing-extensions = "^4.12.2"
werkzeug = "^3.0.4"
prometheus-flask-exporter = "^0.23.1"
pytest-cov = "5.0.0"

[tool.poetry.group.dev.dependencies]
mongomock = "^4.2.0.post1"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"